refactor(overlay): content prop + Tooltip-consistent API
Align with Tooltip/HoverCard pattern: children = base, content = overlay.

- XDSOverlay: content prop replaces XDSOverlayScrim children composition
- useXDSOverlay: hook returns containerRef, containerProps, element
- Remove XDSOverlayScrim (public) — replaced by internal OverlayScrim
- Remove OverlayContext — hook manages isOpen directly, no context needed
  between component and scrim since scrim is now rendered internally
- Rename useXDSOverlay.ts → .tsx (contains JSX)

API:
  <XDSOverlay showOn="hover" content={<Button />}>
    <img src={src} />
  </XDSOverlay>

Hook:
  const overlay = useXDSOverlay({ showOn: "hover", content: <Button /> });
  <Card ref={overlay.containerRef} {...overlay.containerProps}>
    {overlay.element}
  </Card>
