+++ apps/storybook/stories/Overlay.stories.tsx
@@ -1,12 +1,7 @@
 import type {Meta, StoryObj} from '@storybook/react';
 import React, {useState} from 'react';
 import * as stylex from '@stylexjs/stylex';
-import {
-  XDSOverlay,
-  XDSOverlayScrim,
-  useXDSOverlay,
-  OverlayContext,
-} from '@xds/core/Overlay';
+import {XDSOverlay, useXDSOverlay} from '@xds/core/Overlay';
 import {XDSAspectRatio} from '@xds/core/AspectRatio';
 import {XDSCard} from '@xds/core/Card';
 import {XDSButton} from '@xds/core/Button';
@@ -19,10 +14,6 @@ import {
   spacingVars,
   typographyVars,
 } from '@xds/core/theme/tokens.stylex';
-
-// =============================================================================
-// Styles
-// =============================================================================
 
 const styles = stylex.create({
   pageWrapper: {
@@ -71,13 +62,9 @@ const SAMPLE_IMAGE_2 = 'https://picsum.photos/seed/xds-overlay-2/800/450';
 const SAMPLE_IMAGE_3 = 'https://picsum.photos/seed/xds-overlay-3/800/450';
 const SAMPLE_HERO = 'https://picsum.photos/seed/xds-hero/1200/600';
 
-// =============================================================================
-// Meta
-// =============================================================================
-
-const meta: Meta<typeof XDSOverlayScrim> = {
+const meta: Meta<typeof XDSOverlay> = {
   title: 'Core/Overlay',
-  component: XDSOverlayScrim,
+  component: XDSOverlay,
   tags: ['autodocs'],
   decorators: [
     Story => (
@@ -86,29 +73,19 @@ const meta: Meta<typeof XDSOverlayScrim> = {
       </div>
     ),
   ],
-  argTypes: {
-    showOn: {
-      control: 'select',
-      options: ['always', 'hover', 'focus', 'hover-or-focus'],
-    },
-    scrim: {control: 'select', options: ['dark', 'light', false]},
-    position: {control: 'select', options: ['fill', 'bottom', 'top']},
-    align: {control: 'select', options: ['start', 'center', 'end']},
-  },
 };
 
 export default meta;
-type Story = StoryObj<typeof XDSOverlayScrim>;
-
-// =============================================================================
-// Stories
-// =============================================================================
+type Story = StoryObj<typeof XDSOverlay>;
 
 /** Basic hover overlay on an image. */
 export const HoverOnImage: Story = {
   render: () => (
     <div style={{width: 400}}>
-      <XDSOverlay>
+      <XDSOverlay
+        showOn="hover"
+        align="center"
+        content={<XDSButton label="Quick view" variant="ghost" />}>
         <XDSAspectRatio ratio={16 / 9}>
           <img
             src={SAMPLE_IMAGE}
@@ -116,9 +93,6 @@ export const HoverOnImage: Story = {
             {...stylex.props(styles.image)}
           />
         </XDSAspectRatio>
-        <XDSOverlayScrim showOn="hover" align="center">
-          <XDSButton label="Quick view" variant="ghost" />
-        </XDSOverlayScrim>
       </XDSOverlay>
     </div>
   ),
@@ -128,27 +102,32 @@ export const HoverOnImage: Story = {
 export const BottomStrip: Story = {
   render: () => (
     <div style={{width: 600}}>
-      <XDSOverlay>
-        <XDSAspectRatio ratio={2}>
-          <img src={SAMPLE_HERO} alt="Hero" {...stylex.props(styles.image)} />
-        </XDSAspectRatio>
-        <XDSOverlayScrim position="bottom" align="start">
+      <XDSOverlay
+        position="bottom"
+        align="start"
+        content={
           <XDSVStack gap={1}>
             <XDSHeading level={3}>Gallery Collection</XDSHeading>
             <XDSText type="supporting" color="secondary">
               24 items · Updated today
             </XDSText>
           </XDSVStack>
-        </XDSOverlayScrim>
+        }>
+        <XDSAspectRatio ratio={2}>
+          <img src={SAMPLE_HERO} alt="Hero" {...stylex.props(styles.image)} />
+        </XDSAspectRatio>
       </XDSOverlay>
     </div>
   ),
 };
 
-/** Full overlay on Card — uses XDSOverlay wrapper. */
+/** Full overlay wrapping a Card. */
 export const CardOverlay: Story = {
   render: () => (
-    <XDSOverlay>
+    <XDSOverlay
+      showOn="hover"
+      align="center"
+      content={<XDSButton label="View Details" variant="ghost" />}>
       <XDSCard width={360}>
         <XDSLayout
           content={
@@ -167,18 +146,18 @@ export const CardOverlay: Story = {
           }
         />
       </XDSCard>
-      <XDSOverlayScrim showOn="hover" align="center">
-        <XDSButton label="View Details" variant="ghost" />
-      </XDSOverlayScrim>
     </XDSOverlay>
   ),
 };
 
-/** Hover + focus overlay — also appears on keyboard focus. */
+/** Hover + focus — also appears on keyboard focus. */
 export const HoverOrFocus: Story = {
   render: () => (
     <div style={{width: 400}}>
-      <XDSOverlay>
+      <XDSOverlay
+        showOn="hover-or-focus"
+        align="center"
+        content={<XDSButton label="Edit" variant="ghost" />}>
         <XDSAspectRatio ratio={16 / 9}>
           <img
             src={SAMPLE_IMAGE_2}
@@ -186,9 +165,6 @@ export const HoverOrFocus: Story = {
             {...stylex.props(styles.image)}
           />
         </XDSAspectRatio>
-        <XDSOverlayScrim showOn="hover-or-focus" align="center">
-          <XDSButton label="Edit" variant="ghost" />
-        </XDSOverlayScrim>
       </XDSOverlay>
     </div>
   ),
@@ -204,7 +180,16 @@ export const LoadingOverlay: Story = {
           label={isUploading ? 'Cancel upload' : 'Simulate upload'}
           onClick={() => setIsUploading(v => !v)}
         />
-        <XDSOverlay>
+        <XDSOverlay
+          isOpen={isUploading}
+          scrim="light"
+          align="center"
+          content={
+            <XDSVStack gap={2} hAlign="center">
+              <XDSSpinner size="md" />
+              <XDSText weight="bold">Uploading...</XDSText>
+            </XDSVStack>
+          }>
           <XDSAspectRatio ratio={1}>
             <img
               src={SAMPLE_IMAGE_3}
@@ -212,12 +197,6 @@ export const LoadingOverlay: Story = {
               {...stylex.props(styles.image)}
             />
           </XDSAspectRatio>
-          <XDSOverlayScrim isOpen={isUploading} scrim="light" align="center">
-            <XDSVStack gap={2} hAlign="center">
-              <XDSSpinner size="md" />
-              <XDSText weight="bold">Uploading...</XDSText>
-            </XDSVStack>
-          </XDSOverlayScrim>
         </XDSOverlay>
       </XDSVStack>
     );
@@ -238,7 +217,12 @@ export const GalleryGrid: Story = {
     return (
       <XDSGrid columns={3} gap={4}>
         {images.map(img => (
-          <XDSOverlay key={img.title}>
+          <XDSOverlay
+            key={img.title}
+            showOn="hover"
+            position="bottom"
+            align="start"
+            content={<XDSText weight="bold">{img.title}</XDSText>}>
             <XDSAspectRatio ratio={1}>
               <img
                 src={img.src}
@@ -246,9 +230,6 @@ export const GalleryGrid: Story = {
                 {...stylex.props(styles.image)}
               />
             </XDSAspectRatio>
-            <XDSOverlayScrim showOn="hover" position="bottom" align="start">
-              <XDSText weight="bold">{img.title}</XDSText>
-            </XDSOverlayScrim>
           </XDSOverlay>
         ))}
       </XDSGrid>
@@ -256,68 +237,63 @@ export const GalleryGrid: Story = {
   },
 };
 
-/** Video thumbnail with two overlays: duration badge + hover play button. */
+/** Video thumbnail with duration badge (always visible) + hover play overlay. */
 export const VideoThumbnail: Story = {
   render: () => (
     <div style={{width: 400}}>
-      <XDSOverlay>
-        <XDSAspectRatio ratio={16 / 9}>
-          <img src={SAMPLE_IMAGE} alt="Video" {...stylex.props(styles.image)} />
-        </XDSAspectRatio>
-        <XDSOverlayScrim
-          showOn="always"
-          position="bottom"
-          scrim={false}
-          align="end">
-          <div {...stylex.props(styles.durationBadge)}>
-            <XDSText type="supporting">12:34</XDSText>
-          </div>
-        </XDSOverlayScrim>
-        <XDSOverlayScrim showOn="hover" align="center">
+      <XDSOverlay
+        showOn="hover"
+        align="center"
+        content={
           <XDSVStack gap={2} hAlign="center">
             <XDSText weight="bold" size="lg">
               ▶
             </XDSText>
             <XDSText weight="bold">Introduction to XDS</XDSText>
           </XDSVStack>
-        </XDSOverlayScrim>
+        }>
+        <XDSAspectRatio ratio={16 / 9}>
+          <img src={SAMPLE_IMAGE} alt="Video" {...stylex.props(styles.image)} />
+        </XDSAspectRatio>
       </XDSOverlay>
     </div>
   ),
 };
 
-/** Disconnected hover — uses useXDSOverlay hook on Card. Hovering anywhere on Card reveals overlay on image. */
+/** Disconnected hover — useXDSOverlay hook on Card. Hover Card reveals overlay on image. */
 export const DisconnectedHover: Story = {
   render: function DisconnectedHoverStory() {
-    const overlay = useXDSOverlay();
+    const overlay = useXDSOverlay({
+      showOn: 'hover',
+      position: 'bottom',
+      align: 'start',
+      content: <XDSHeading level={4}>Featured Article</XDSHeading>,
+    });
+
     return (
-      <OverlayContext value={overlay.contextValue}>
-        <XDSCard
-          width={360}
-          ref={overlay.containerRef as React.RefObject<HTMLDivElement>}
-          {...overlay.containerProps}>
-          <div {...stylex.props(styles.imageSection)}>
-            <XDSAspectRatio ratio={16 / 9}>
-              <img
-                src={SAMPLE_IMAGE_2}
-                alt="Article"
-                {...stylex.props(styles.image)}
-              />
-            </XDSAspectRatio>
-            <XDSOverlayScrim showOn="hover" position="bottom" align="start">
-              <XDSHeading level={4}>Featured Article</XDSHeading>
-            </XDSOverlayScrim>
-          </div>
-          <XDSVStack gap={1} xstyle={styles.metadata}>
-            <XDSText type="supporting" color="secondary">
-              Jan 15, 2026 · 5 min read
-            </XDSText>
-            <XDSText type="supporting" color="secondary">
-              By Jane Author
-            </XDSText>
-          </XDSVStack>
-        </XDSCard>
-      </OverlayContext>
+      <XDSCard
+        width={360}
+        ref={overlay.containerRef as React.RefObject<HTMLDivElement>}
+        {...overlay.containerProps}>
+        <div {...stylex.props(styles.imageSection)}>
+          <XDSAspectRatio ratio={16 / 9}>
+            <img
+              src={SAMPLE_IMAGE_2}
+              alt="Article"
+              {...stylex.props(styles.image)}
+            />
+          </XDSAspectRatio>
+          {overlay.element}
+        </div>
+        <XDSVStack gap={1} xstyle={styles.metadata}>
+          <XDSText type="supporting" color="secondary">
+            Jan 15, 2026 · 5 min read
+          </XDSText>
+          <XDSText type="supporting" color="secondary">
+            By Jane Author
+          </XDSText>
+        </XDSVStack>
+      </XDSCard>
     );
   },
 };
@@ -333,18 +309,20 @@ export const DragAndDrop: Story = {
           variant="secondary"
           onClick={() => setIsDragOver(v => !v)}
         />
-        <XDSOverlay>
+        <XDSOverlay
+          isOpen={isDragOver}
+          align="center"
+          content={
+            <XDSVStack gap={2} hAlign="center">
+              <XDSText size="lg">📁</XDSText>
+              <XDSText weight="bold">Drop files here</XDSText>
+            </XDSVStack>
+          }>
           <div {...stylex.props(styles.dropZone)}>
             <p {...stylex.props(styles.dropZoneText)}>
               Drop files here or click to browse
             </p>
           </div>
-          <XDSOverlayScrim isOpen={isDragOver} scrim="dark" align="center">
-            <XDSVStack gap={2} hAlign="center">
-              <XDSText size="lg">📁</XDSText>
-              <XDSText weight="bold">Drop files here</XDSText>
-            </XDSVStack>
-          </XDSOverlayScrim>
         </XDSOverlay>
       </XDSVStack>
     );
@@ -355,7 +333,11 @@ export const DragAndDrop: Story = {
 export const NoScrim: Story = {
   render: () => (
     <div style={{width: 300}}>
-      <XDSOverlay>
+      <XDSOverlay
+        showOn="hover"
+        scrim={false}
+        align="center"
+        content={<XDSButton label="♡" variant="ghost" />}>
         <XDSAspectRatio ratio={1}>
           <img
             src={SAMPLE_IMAGE_3}
@@ -363,9 +345,6 @@ export const NoScrim: Story = {
             {...stylex.props(styles.image)}
           />
         </XDSAspectRatio>
-        <XDSOverlayScrim showOn="hover" scrim={false} align="center">
-          <XDSButton label="♡" variant="ghost" />
-        </XDSOverlayScrim>
       </XDSOverlay>
     </div>
   ),
