+++ packages/core/src/Overlay/index.ts
@@ -6,17 +6,12 @@
 export {XDSOverlay} from './XDSOverlay';
 export type {XDSOverlayProps} from './XDSOverlay';
 
-export {XDSOverlayScrim} from './XDSOverlayScrim';
+export {useXDSOverlay} from './useXDSOverlay';
+export type {UseXDSOverlayOptions, UseXDSOverlayResult} from './useXDSOverlay';
+
 export type {
-  XDSOverlayScrimProps,
   OverlayScrimMode,
   OverlayPosition,
   OverlayAlign,
   OverlayShowOn,
-} from './XDSOverlayScrim';
-
-export {useXDSOverlay} from './useXDSOverlay';
-export type {UseXDSOverlayOptions, UseXDSOverlayResult} from './useXDSOverlay';
-
-export {OverlayContext} from './OverlayContext';
-export type {OverlayContextValue} from './OverlayContext';
+} from './OverlayScrim';
